revert change of refresh token rotation to "should"
this reverts the change in -07 that changed refresh token rotation from "must" to "should". Instead, it now matches the Security BCP recommendation of MUST either do rotation or sender-constrained refresh tokens. I was unable to track down the source that motivated that change from must to should, I couldn't find references in the meeting minutes or on the mailing list.