Address #80
Clarified that we only focus on OAuth and not on other consequences (which are more broad and apply to all types of JS applications, regardless of OAuth).

Hope this suffices, especially because the attack scenarios are clearly marked as non-exhaustive. The consequences are pretty well-defined and scoped, so I think this should work.